+++ motor-control/timer.h
@@ -11,7 +11,7 @@
 #define MS_TIMER    	TCC0
 #define MS_TIMER_PER	5		// Period of MS_TIMER in milliseconds
 
-extern volatile uint8_t ms_timer;
+extern volatile uint16_t ms_timer;
 
 void init_pwm_timer(TC0_t *timer);
 void init_enc_timer(TC1_t *timer, TC_EVSEL_t event_channel);
